requirements.txt
flask==2.3.3
openai==1.3.0
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.14.1
openpyxl
groq